backend/pyproject.toml
[tool.poetry]
name = "energy_dashboard"
version = "0.1.0"
description = ""
authors = ["Zahyr Seferina <zahyr.seferina@gmail.com>"]

[tool.poetry.scripts]
start = "energy_dashboard.main:start"

[tool.poetry.dependencies]
python = "^3.12"
websockets = "^12.0"
uvicorn = "^0.30.1"
fastapi = "^0.111.0"
requests = "^2.32.3"
pydantic = "^2.8.0"
python-homewizard-energy = "^10.0.0"
pytest = "^8.2.2"
sqlalchemy = "^2.0.31"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
